Remove netcoreapp SecureString dependency
