+++ source/Calamari/Integration/Processes/CommandLineInvocation.cs
@@ -1,4 +1,3 @@
-using System.Security;
 using Calamari.Util;
 
 namespace Calamari.Integration.Processes
@@ -13,7 +12,7 @@ namespace Calamari.Integration.Processes
             Arguments = arguments;
         }
 
-        public CommandLineInvocation(string executable, string arguments, string workingDirectory, string userName = null, SecureString password = null)
+        public CommandLineInvocation(string executable, string arguments, string workingDirectory, string userName = null, string password = null)
             : this(executable, arguments)
         {
             this.workingDirectory = workingDirectory;
@@ -27,7 +26,7 @@ namespace Calamari.Integration.Processes
 
         public string UserName { get; }
 
-        public SecureString Password { get; }
+        public string Password { get; }
 
         /// <summary>
         /// The initial working-directory for the invocation.
