+++ source/Calamari/project.json
@@ -80,7 +80,6 @@
         "System.Runtime.Extensions": "4.1.0",
         "System.Runtime.InteropServices.RuntimeInformation": "4.0.0",
         "System.Runtime.Serialization.Formatters": "4.0.0-rc3-24212-01",
-        "System.Security.SecureString": "4.0.0",
         "System.Security.Cryptography.Primitives": "4.0.0",
         "System.Security.Cryptography.X509Certificates": "4.1.0",
         "System.Threading.AccessControl": "4.0.0",
